Moved Renderable to a separate file
