+++ src/render-system.cpp
@@ -8,21 +8,12 @@
 #include "graphics/shader.hpp"
 #include "graphics/vertex-buffer-object.hpp"
 #include "components/transforms.hpp"
+#include "components/renderable.hpp"
 #include "graphics/material.hpp"
 #include "entity.hpp"
 #include "os.hpp"
 
 namespace tec {
-	Renderable::Renderable(std::shared_ptr<VertexBufferObject> buf,
-		std::shared_ptr<Shader> shader) : buffer(buf), shader(shader) {
-		if (this->buffer) {
-			size_t group_count = this->buffer->GetVertexGroupCount();
-			for (size_t i = 0; i < group_count; ++i) {
-				this->vertex_groups.insert(this->buffer->GetVertexGroup(i));
-			}
-		}
-	}
-
 	RenderSystem::RenderSystem() : window_width(800), window_height(600) {
 		auto err = glGetError();
 		// If there is an error that means something went wrong when creating the context.
