+++ src/test_data.cpp
@@ -4,14 +4,16 @@
 #include "graphics/texture-object.hpp"
 #include "components/transforms.hpp"
 #include "components/collisionbody.hpp"
+#include "components/renderable.hpp"
 #include "resources/md5mesh.hpp"
 #include "resources/obj.hpp"
 #include "resources/pixel-buffer.hpp"
 #include "resources/md5anim.hpp"
 #include "resources/vorbis-stream.hpp"
+#include "graphics/animation.hpp"
+#include "render-system.hpp"
 #include "entity.hpp"
 #include "component-update-system.hpp"
-#include "render-system.hpp"
 #include "sound-system.hpp"
 #include "physics-system.hpp"
 #include "voxelvolume.hpp"
